backend/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
python-dotenv>=1.0.0
pydantic-settings>=2.2.0

# ML & Cheminformatics
torch>=2.0.0
rdkit>=2023.3.1
backoff>=2.2.1
scikit-learn>=1.4.0
numpy>=1.26.0
pandas>=2.2.0
pyarrow>=15.0.0

# Pretrained Molecular Encoder (ChemBERTa from Hugging Face)
transformers>=4.40.0
sentencepiece>=0.1.99

# LLM & Agents
google-generativeai>=0.5.0
langchain>=0.2.0
langchain-google-genai>=1.0.0
langchain-community>=0.2.0
pyautogen>=0.2.0

# ZINC Dataset + Vector Store
datasets>=2.18.0
faiss-cpu>=1.8.0
